crces-backend/requirements.txt
amqp==5.3.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
pyodbc==5.2.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==8.2
redis==6.4.0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
